+++ src/virtio_pci.zig
@@ -338,7 +338,7 @@ pub fn get_registered_devs() ?*Dev {
     return registered_devs;
 }
 
-fn register_dev(comptime kind: enum { blk, net }, allocator: mem.Allocator, irq: u8, h: *const fn (*Dev, u64, io.Operation, []u8) anyerror!void) !*Dev {
+fn register_dev(comptime kind: enum { blk, net }, allocator: mem.Allocator, irq: u32, h: *const fn (*Dev, u64, io.Operation, []u8) anyerror!void) !*Dev {
     const redhat_qumranet_vendor = 0x1af4;
     const device_id = switch (kind) {
         .blk => 0x1001, // PCI_DEVICE_ID_VIRTIO_BLK
@@ -362,10 +362,10 @@ fn register_dev(comptime kind: enum { blk, net }, allocator: mem.Allocator, irq:
     return vpdev;
 }
 
-pub fn register_blk_dev(allocator: mem.Allocator, irq: u8, h: *const fn (*Dev, u64, io.Operation, []u8) anyerror!void) !*Dev {
+pub fn register_blk_dev(allocator: mem.Allocator, irq: u32, h: *const fn (*Dev, u64, io.Operation, []u8) anyerror!void) !*Dev {
     return register_dev(.blk, allocator, irq, h);
 }
 
-pub fn register_net_dev(allocator: mem.Allocator, irq: u8, h: *const fn (*Dev, u64, io.Operation, []u8) anyerror!void) !*Dev {
+pub fn register_net_dev(allocator: mem.Allocator, irq: u32, h: *const fn (*Dev, u64, io.Operation, []u8) anyerror!void) !*Dev {
     return register_dev(.net, allocator, irq, h);
 }
